feat(bottom-tab): added the correct icon and highlight functionnalities
